+++ packages/sdk/test/ci/module-schema.test.ts
@@ -13,7 +13,7 @@ describe('VersionedModuleSource: getEmbeddedModuleSchema', () => {
         );
         const moduleSource = versionedModuleSourceFromBuffer(contractModule);
         const moduleSchema = await getEmbeddedModuleSchema(moduleSource);
-        if (moduleSchema === null) {
+        if (moduleSchema === undefined) {
             fail('Failed to find module schame');
         }
         expect(moduleSchema.type).toBe('versioned');
@@ -25,7 +25,7 @@ describe('VersionedModuleSource: getEmbeddedModuleSchema', () => {
         );
         const moduleSource = versionedModuleSourceFromBuffer(contractModule);
         const moduleSchema = await getEmbeddedModuleSchema(moduleSource);
-        if (moduleSchema === null) {
+        if (moduleSchema === undefined) {
             fail('Failed to find module schame');
         }
         expect(moduleSchema.type).toBe('versioned');
@@ -40,7 +40,7 @@ describe('VersionedModuleSource: getEmbeddedModuleSchema', () => {
             source: Buffer.from(unversionedContractModule),
         } as const;
         const moduleSchema = await getEmbeddedModuleSchema(moduleSource);
-        if (moduleSchema === null) {
+        if (moduleSchema === undefined) {
             fail('Failed to find module schame');
         }
         expect(moduleSchema.type).toBe('unversioned');
@@ -52,7 +52,7 @@ describe('VersionedModuleSource: getEmbeddedModuleSchema', () => {
         );
         const moduleSource = versionedModuleSourceFromBuffer(contractModule);
         const moduleSchema = await getEmbeddedModuleSchema(moduleSource);
-        if (moduleSchema === null) {
+        if (moduleSchema === undefined) {
             fail('Failed to find module schame');
         }
         expect(moduleSchema.type).toBe('unversioned');